refactor shared components
