+++ apps/dashboard/src/app/app.module.ts
@@ -6,13 +6,14 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ComponentsModule } from './components/components.module';
+import { SharedComponentsModule } from './components/components.module';
 
 import * as P from './pages';
 import { MaterialComponentsModule } from './components/material-components.module';
 import { AppDirectivesModule } from './directives';
+import { AppTablesModule } from './pages/tables/tables.module';
 
-const Pages = [P.ConnectionComponent, P.ExportComponent, P.TablesComponent, P.UsersComponent];
+const Pages = [P.ConnectionComponent, P.ExportComponent, P.UsersComponent];
 
 @NgModule({
   declarations: [AppComponent, ...Pages],
@@ -21,11 +22,12 @@ const Pages = [P.ConnectionComponent, P.ExportComponent, P.TablesComponent, P.Us
     BrowserAnimationsModule,
     FormsModule,
     ReactiveFormsModule,
-    ComponentsModule,
+    SharedComponentsModule,
     AppRoutingModule,
     MaterialComponentsModule,
     AppDirectivesModule,
     HttpClientModule,
+    AppTablesModule,
   ],
   providers: [],
   bootstrap: [AppComponent],
