+++ apps/dashboard/src/app/pages/tables/components/table-row-editor/components/input-attributes-display.ts
