+++ apps/dashboard/src/app/pages/tables/components/table-row-editor/prompts/cwbc/custom_section_complete.ts
