+++ apps/dashboard/src/app/pages/tables/components/table-row-editor/prompts/directives/input-attributes.directive.ts
