+++ apps/dashboard/src/app/pages/tables/components/table-row-editor/prompts/prompts.scss
