+++ apps/dashboard/src/app/pages/tables/components/table-row-editor/table-row-editor.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MaterialComponentsModule } from '../material-components.module';
+import { MaterialComponentsModule } from '../../../../components/material-components.module';
 import { TableRowEditorDialogComponent } from './table-row-editor';
 import { ODKXPromptsModule } from './prompts/prompts.module';
 import { InputAttributesDisplay } from './components/input-attributes-display';
-import { SharedPipesModule } from '../../pipes';
+import { SharedPipesModule } from '../../../../pipes';
 
 @NgModule({
   imports: [
