+++ apps/dashboard/src/app/pages/tables/components/table-row-editor/table-row-editor.scss
