+++ apps/dashboard/src/environments/environment.prod.ts
@@ -1,8 +1,9 @@
 import DASHBOARD_SETTINGS from '../assets/dashboardSettings.json';
+import {version} from '../../../../package.json'
 
 export const environment = {
   production: true,
-  appVersion: require('../../../../package.json').version,
+  appVersion: version,
   useApiProxy: true,
   ...DASHBOARD_SETTINGS,
 };
