+++ apps/dashboard/src/environments/environment.ts
@@ -1,8 +1,9 @@
 import DASHBOARD_SETTINGS from '../assets/dashboardSettings.json';
+import {version} from '../../../../package.json'
 
 export const environment = {
   production: false,
-  appVersion: require('../../../../package.json').version,
+  appVersion: version,
   /** Proxy all requests to /api to be intercepted by local proxy (to avoid cors issues) */
   useApiProxy: true,
   ...DASHBOARD_SETTINGS,
