+++ tsconfig.base.json
@@ -10,7 +10,7 @@
     "importHelpers": true,
     "target": "es2015",
     "module": "es2020",
-    "typeRoots": ["node_modules/@types"],
+    "typeRoots": ["./node_modules/@types"],
     "lib": ["es2017", "dom"],
     "skipLibCheck": true,
     "skipDefaultLibCheck": true,
